// apps/nestjs/src/modules/media/tiktok/interface/tik-report.interface.ts
export interface TiktokReport {
  advertiserId: number;
  adId: number;
  statTimeDay: string;
  budget: number;
  spend: number;
  impressions: number;
  clicks: number;
  videoPlayActions: number;
  videoWatched2s: number;
  videoWatched6s: number;
  videoViewsP100: number;
  reach: number;
  conversion: number;
  campaignId: number;
  campaignName: string;
  adgroupId: number;
  adgroupName: string;
  adName: string;
  adUrl: string;
  statTimeDayDimension: string;
  adIdDimension: string;
}